Fix unit test failed
