+++ core/testing/src/main/kotlin/com/purkt/mindexpense/core/testing/base/BaseTest.kt
@@ -1,12 +1,6 @@
 package com.purkt.mindexpense.core.testing.base
 
-import com.purkt.mindexpense.core.logging.AppLogger
-import io.mockk.every
-import io.mockk.just
 import io.mockk.mockkClass
-import io.mockk.mockkObject
-import io.mockk.runs
-import io.mockk.unmockkObject
 import org.junit.Rule
 import org.koin.dsl.ModuleDeclaration
 import org.koin.dsl.module
@@ -36,20 +30,5 @@ abstract class BaseTest: KoinTest {
     @get:Rule
     val mockProvider = MockProviderRule.create { clazz ->
         mockkClass(clazz)
-    }
-
-    companion object {
-        fun setUpMockAppLogger() {
-            mockkObject(AppLogger)
-            every { AppLogger.i(message = any()) } just runs
-            every { AppLogger.d(message = any()) } just runs
-            every { AppLogger.v(message = any()) } just runs
-            every { AppLogger.w(message = any()) } just runs
-            every { AppLogger.e(message = any()) } just runs
-        }
-
-        fun tearDownMockAppLogger() {
-            unmockkObject(AppLogger)
-        }
     }
 }
